requirements.txt
discord.py == 1.7.3
discord == 1.7.3
requests == 2.26.0
jikanpy == 4.3.2
bs4 == 0.0.1
lxml == 4.6.4
pyaml == 21.10.1